RG-00 add student to course functionality

+++ adminservice/src/main/java/com/myschool/adminservice/controller/MessageController.java
@@ -36,7 +36,7 @@ public class MessageController {
     }
 
     @GetMapping(value = "allcoursemessages/{courseId}")
-    public List<CourseMessage> getMessagesByCourseId(@PathVariable(value = "courseId") long courseId) {
+    public List<CourseMessage> getMessagesByCourseId(@PathVariable(value = "courseId") Integer courseId) {
         List<CourseMessage> courseMessageList = messageService.getCourseMessagesByCourseId(courseId);
         return courseMessageList;
     }
