+++ adminservice/src/main/java/com/myschool/adminservice/repository/CourseWorkRepository.java
@@ -10,5 +10,5 @@ import java.util.List;
 public interface CourseWorkRepository extends PagingAndSortingRepository<CourseWork, Long> {
     //Return everything except uploadedfile
     @Query("SELECT new CourseWork(c.id, c.type, c.title, c.description, c.topic, c.createdTimeStamp, c.videoLink,c.fileName, c.fileSize, c.courseId) FROM CourseWork c where c.courseId = :courseId")
-    List<CourseWork> findAllByCourseId(long courseId, Sort sort);
+    List<CourseWork> findAllByCourseId(Integer courseId, Sort sort);
 }
